Update the output json

+++ eval/src/sort.rs
@@ -36,30 +36,18 @@ fn compute_next_count(zero_flag: bool, next_count: usize) -> usize {
 }
 
 #[cap]
-fn sel1(cond: bool, a: u32, b: u32) -> u32 {
-    if cond {
-        a
-    } else {
-        b
-    }
+fn sel1(cond: bool, a: usize, b: usize) -> usize {
+    if cond { a } else { b }
 }
 
 #[cap]
-fn sel2(cond: bool, a: u32, b: u32) -> u32 {
-    if cond {
-        a
-    } else {
-        b
-    }
+fn sel2(cond: bool, a: usize, b: usize) -> usize {
+    if cond { a } else { b }
 }
 
 #[cap]
-fn sel3(cond: bool, a: u32, b: u32) -> u32 {
-    if cond {
-        a
-    } else {
-        b
-    }
+fn sel3(cond: bool, a: usize, b: usize) -> usize {
+    if cond { a } else { b }
 }
 
 #[cap(a: uniq @ 0..N, z: uniq @ 0..N)]
